Further schema definition. (#16)
Main changes:
- Adds the majority of needed checks to schema, excluding threshhold
checks*.
- Switches to raising warnings from early pipeline schema.
- Limits the failure cases returned in error messages.
- Sets primary keys with `pandera.typing.Index` and multiindex configs.
- Adds missing columns and filters extra columns/indices in early
pipeline schema.
- Sets categorical types (in addition to existing enum coercion).
- Adds missing bound to data definition.

Secondary changes:
- Switches to importing checks via their submodules.
- Switches to importing constants module instead of each constant.
- Comments planned further normalization.
- Adds to list of needed validations within pipeline.
- Adds a sample image of a datasheet to `assets/`.
- Reorders some fields.
- Adds/adjusts TODOs and comments.
- Updates `safety` pin to incorporate patch.

NOTE: This is still untested. Will likely need to finagle as we try it,
and definitely will need to build test coverage of these definitions,
constraints, and validations.

* Thresholds depend on the type of the site, outfall or creek, and we
don't have a table for that just yet. We can make a dummy table and
implement the logic, but I left that for a later PR since this one is
getting big enough and I've been sitting on it for a while.